# app/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
...
        "/$modelName$/listByRole": {
            "post": {
                "description": "\"ListByRole\"",
                "tags": [
                    "operation"
                ],
                "parameters": [
                    {
                        "description": "data",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/api.paramQueryRole"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "#/definitions/$modelName$",
                        "schema": {
                            "type": "array",
                            "items": {
                                "type": "object"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/i18n.CustomError"
                        }
                    }
                }
            }
        },
        "/$modelName$/listPublic": {
            "get": {
                "description": "\"ListPublic\"",
                "tags": [
                    "operation"
                ],
                "responses": {
                    "200": {
                        "description": "#/definitions/$modelName$",
                        "schema": {
                            "type": "array",
                            "items": {
                                "type": "object"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/i18n.CustomError"
                        }
                    }
                }
            }
        },
...
        "/globalOperation/authenticationHookOptions": {
            "get": {
                "description": "\"authenticationHookOptions\"",
                "tags": [
                    "globalOperation"
                ],
                "responses": {
                    "200": {
                        "description": "authenticationHookOptions配置",
                        "schema": {
                            "$ref": "#/definitions/models.HookOptions"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/i18n.CustomError"
                        }
                    }
                }
            }
        },
        "/globalOperation/httpTransportHookOptions": {
            "get": {
                "description": "\"httpTransportHookOptions\"",
                "tags": [
                    "globalOperation"
                ],
                "responses": {
                    "200": {
                        "description": "httpTransportHookOptions配置",
                        "schema": {
                            "$ref": "#/definitions/models.HookOptions"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/i18n.CustomError"
                        }
                    }
                }
            }
        },
...
        "/operation/bindRoles": {
            "post": {
                "description": "\"BindRoles\"",
                "tags": [
                    "operation"
                ],
                "parameters": [
                    {
                        "description": "data",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/api.paramBindRole"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "type": "string"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/i18n.CustomError"
                        }
                    }
                }
            }
        },
        "/operation/function/{dataName}": {
            "post": {
                "description": "\"updateFunctionText\"",
                "tags": [
                    "operation"
                ],
                "parameters": [
                    {
                        "type": "string",
                        "description": "dataName",
                        "name": "dataName",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "文本",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/i18n.CustomError"
                        }
                    }
                }
            }
        },
        "/operation/graphql/{dataName}": {
            "get": {
                "description": "\"GetGraphqlText\"",
                "tags": [
                    "operation"
                ],
                "parameters": [
                    {
                        "type": "string",
                        "description": "dataName",
                        "name": "dataName",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "post": {
                "description": "\"UpdateGraphqlText\"",
                "tags": [
                    "operation"
                ],
                "parameters": [
                    {
                        "type": "string",
                        "description": "dataName",
                        "name": "dataName",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "文本",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/i18n.CustomError"
                        }
                    }
                }
            }
        },
        "/operation/hookOptions/{dataName}": {
            "get": {
                "description": "\"getHookOptions\"",
                "tags": [
                    "operation"
                ],
                "parameters": [
                    {
                        "type": "string",
                        "description": "dataName",
                        "name": "dataName",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "hook配置",
                        "schema": {
                            "$ref": "#/definitions/models.HookOptions"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/i18n.CustomError"
                        }
                    }
                }
            }
        },
        "/operation/proxy/{dataName}": {
            "post": {
                "description": "\"updateProxyText\"",
                "tags": [
                    "operation"
                ],
                "parameters": [
                    {
                        "type": "string",
                        "description": "dataName",
                        "name": "dataName",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "文本",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/i18n.CustomError"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "api.paramBindRole": {
            "type": "object",
            "properties": {
                "operationPaths": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "rbacType": {
                    "type": "string"
                },
                "roleCodes": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                }
            }
        },
        "api.paramQueryRole": {
            "type": "object",
            "properties": {
                "rbacType": {
                    "type": "string"
                },
                "roleCode": {
                    "type": "string"
                }
            }
        },
...
        "i18n.CustomError": {
            "type": "object",
            "properties": {
                "code": {
                    "description": "错误码",
                    "allOf": [
                        {
                            "$ref": "#/definitions/i18n.Errcode"
                        }
                    ]
                },
                "message": {
                    "description": "错误消息",
                    "type": "string"
                },
                "mode": {
                    "description": "业务模块",
                    "type": "string"
                }
            }
        },
        "i18n.Errcode": {
            "type": "integer",
            "enum": [
                10101,
                10201,
                10202,
                10203,
                10204,
                10205,
                10206,
                10207,
                10301,
                10302,
                10303,
                10304,
                10305,
                10401,
                10402,
                10403,
                10404,
                10405,
                10406,
                10407,
                10501,
                10502,
                10503,
                10504,
                10505,
                10506,
                10507,
                10508,
                10509,
                10510,
                10511,
                10512,
                10513,
                10514,
                10515,
                10516,
                10517,
                10518,
                10519,
                10520,
                10601,
                10602,
                10603,
                10604,
                10605,
                10606,
                20101,
                10102,
                20201,
                20202,
                20203,
                20204,
                20205,
                20206,
                20207,
                20208,
                20209,
                20210,
                20211,
                20301,
                20302,
                20303,
                20304,
                20305,
                20306,
                20307,
                20401,
                20402,
                20403,
                20404,
                20405,
                20406,
                20407,
                20408,
                20409,
                20501,
                20502,
                20601,
                20701
            ],
            "x-enum-varnames": [
                "ServerError",
                "ParamIllegalError",
                "ParamBindError",
                "StructParamEmtpyError",
                "BodyParamEmptyError",
                "PathParamEmptyError",
                "QueryParamEmptyError",
                "FormParamEmptyError",
                "RequestResubmitError",
                "RequestSignatureError",
                "RequestReadBodyError",
                "RequestEmptyBodyError",
                "RequestProxyError",
                "FileReadError",
                "FileWriteError",
                "FileZipError",
                "FileUnZipError",
                "FileZipAmountZeroError",
                "FileContentEmptyError",
                "DirectoryReadError",
                "LoaderFileReadError",
                "LoaderFileUnmarshalError",
                "LoaderDataExistEditorError",
                "LoaderDataExistError",
                "LoaderDataNotExistError",
                "LoaderLockNotFoundError",
                "LoaderWatcherNotSupport",
                "LoaderNoneModifiedError",
                "LoaderRWNotSupportError",
                "LoaderNameEmptyError",
                "LoaderBasenameEmptyErr",
                "LoaderRootOrExtensionEmptyErr",
                "LoaderMultipleOnlyError",
                "LoaderEmbedNotAllowModifyErr",
                "LoaderRemoveKeyNotFoundError",
                "LoaderRenameKeyNotFoundError",
                "LoaderRenameTargetExistError",
                "LoaderRenameNotAllowMultipleError",
                "LoaderWriteableRelyModelRequiredError",
                "LoaderDataFilepathError",
                "VscodeOnlyDirectoriesCanWatchError",
                "VscodeDirectoryExistError",
                "VscodeFileExistError",
                "VscodeFileNotExistError",
                "VscodeSourceNotDirectoryError",
                "VscodeTargetDirectoryExistError",
                "EngineCreateConfigError",
                "EngineRestartError",
                "DataInsertError",
                "DataDeleteError",
                "DataUpdateError",
                "DataSelectError",
                "DataCopyError",
                "DataRenameError",
                "DataBatchInsertError",
                "DataBatchDeleteError",
                "DataBatchUpdateError",
                "DataEmptyListError",
                "DataNotExistsError",
                "DatasourceConnectionError",
                "DatasourceKindNotSupportedError",
                "DatasourceDisabledError",
                "DatasourceDatabaseUrlEmptyError",
                "DatabaseOasVersionError",
                "PrismaQueryError",
                "PrismaMigrateError",
                "StoragePingError",
                "StorageDisabledError",
                "StorageMkdirError",
                "StorageTouchError",
                "StorageRemoveError",
                "StorageRenameError",
                "StorageListError",
                "StorageDetailError",
                "StorageDownloadError",
                "OperationRoleHasBindError",
                "OperationRbacTypeError",
                "SettingServerUrlEmptyError",
                "SdkAlreadyUpToDateError"
            ]
        },
        "models.HookOptions": {
            "type": "object",
            "additionalProperties": {
                "$ref": "#/definitions/models.hookOption"
            }
        },
...
        "models.hookOption": {
            "type": "object",
            "properties": {
                "enabled": {
                    "type": "boolean"
                },
                "existed": {
                    "type": "boolean"
                },
                "path": {
                    "type": "string"
                }
            }
        },
...
    }
}